% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_puzzle_input.R
\name{get_puzzle_input}
\alias{get_puzzle_input}
\title{Get puzzle input}
\usage{
get_puzzle_input(
  year,
  day,
  cache = TRUE,
  AOC_URL = "https://adventofcode.com/"
)
}
\description{
Get puzzle input for a given year and day
}
